# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Team360 Health Check Application REST API. Assesses team health across 11 dimensions with hierarchical org support.",
        "title": "Team360 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/sso/callback": {
            "post": {
                "description": "Exchanges an authorization code + PKCE verifier for the OAuth provider's tokens, resolves the user by email, and issues our own JWT access/refresh token pair. Public endpoint.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "SSO OAuth callback",
                "parameters": [
                    {
                        "description": "Authorization code and PKCE verifier",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interfaces_api_v1.ssoCallbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Missing code or code_verifier",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Token exchange failed, no account found, or account does not support SSO",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to generate authentication tokens",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "SSO is not configured on this server",
                        "schema": {
                            "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_agopalakrishnan_teams360_backend_interfaces_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_agopalakrishnan_teams360_backend_interfaces_dto.LoginResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "expiresIn": {
                    "description": "Access token expiry in seconds",
                    "type": "integer"
                },
                "refreshToken": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_agopalakrishnan_teams360_backend_interfaces_dto.UserDTO"
                }
            }
        },
...
        "github_com_agopalakrishnan_teams360_backend_interfaces_dto.UserDTO": {
            "type": "object",
            "properties": {
                "canTakeSurvey": {
                    "type": "boolean"
                },
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "hierarchyLevel": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "teamIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "interfaces_api_v1.ssoCallbackRequest": {
            "type": "object",
            "required": [
                "code",
                "code_verifier"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "code_verifier": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}